refactors: Moves ItemDetail and ItemList components to a views folder
