+++ src/components/Items/index.js
@@ -1,6 +1,2 @@
 export * from './ItemCard';
 export * from './ItemCount';
-export * from './ItemDetail';
-export * from './ItemDetailContainer';
-export * from './ItemList';
-export * from './ItemListContainer';
