+++ src/views/ItemDetail/index.js
@@ -1,0 +1,2 @@
+export * from './ItemDetail';
+export * from './ItemDetailContainer';
