+++ src/views/ItemList/index.js
@@ -1,0 +1,2 @@
+export * from './ItemList';
+export * from './ItemListContainer';
